Extracted time input field #22
